+++ spec/serializers/export/aspect_serializer_spec.rb
@@ -6,9 +6,8 @@ describe Export::AspectSerializer do
 
   it "has aspect attributes" do
     expect(serializer.attributes).to eq(
-      name:             aspect.name,
-      contacts_visible: aspect.contacts_visible,
-      chat_enabled:     aspect.chat_enabled
+      name:         aspect.name,
+      chat_enabled: aspect.chat_enabled
     )
   end
 end